backend/requirements.txt
bandit
coverage
django-debug-toolbar
django-debug-toolbar-request-history
django-naomi
isort
prospector
pylint
pylint-celery
safety
pylint-django
model_mommy
pre_commit

Django
celery
django-model-utils
django-webpack-loader
django-js-reverse
django-import-export
python-decouple
brotlipy
django-log-request-id
dj-database-url
gunicorn
whitenoise
psutil
ipython
djangorestframework
dronekit
dronekit_sitl
psycopg2-binary
django-extensions
django-background-tasks
django-nose
geopy
preconditions
Pillow
ExifRead
Shapely
Fiona
numpy
channels
channels-redis
watchdog
django-redis
networkx
utm
cryptography
webp
